move `markMutationResult` and `markMutationOptimistic` from `QueryManager` to `QueryInfo`
