+++ src/core/QueryInfo.ts
@@ -6,15 +6,36 @@ import type { FetchResult } from "@apollo/client/link";
 import type { Unmasked } from "@apollo/client/masking";
 import {
   DeepMerger,
+  getOperationName,
   graphQLResultHasError,
+  isExecutionPatchIncrementalResult,
+  isExecutionPatchResult,
   isNonEmptyArray,
   mergeIncrementalData,
 } from "@apollo/client/utilities/internal";
+import { invariant } from "@apollo/client/utilities/invariant";
+
+import type { IgnoreModifier } from "../cache/core/types/common.js";
 
 import type { ObservableQuery } from "./ObservableQuery.js";
 import type { QueryManager } from "./QueryManager.js";
-import type { OperationVariables } from "./types.js";
-import type { ErrorPolicy, WatchQueryOptions } from "./watchQueryOptions.js";
+import type {
+  DefaultContext,
+  InternalRefetchQueriesInclude,
+  MutationUpdaterFunction,
+  OnQueryUpdated,
+  OperationVariables,
+} from "./types.js";
+import type {
+  ErrorPolicy,
+  MutationFetchPolicy,
+  MutationOptions,
+  WatchQueryOptions,
+} from "./watchQueryOptions.js";
+
+type UpdateQueries<TData> = MutationOptions<TData, any, any>["updateQueries"];
+
+const IGNORE = {} as IgnoreModifier;
 
 export const enum CacheWriteBehavior {
   FORBID,
@@ -64,6 +85,10 @@ export class QueryInfo {
   variables?: Record<string, any>;
 
   private cache: ApolloCache;
+  private queryManager: Pick<
+    QueryManager,
+    "getObservableQueries" | "refetchQueries" | "getDocumentInfo"
+  >;
   public queryId: string;
   // TODO should be private
   public readonly observableQuery?: ObservableQuery<any, any>;
@@ -79,6 +104,7 @@ export class QueryInfo {
     const cache = (this.cache = queryManager.cache);
     this.queryId = queryManager.generateQueryId();
     this.observableQuery = observableQuery;
+    this.queryManager = queryManager;
 
     // Track how often cache.evict is called, since we want eviction to
     // override the feud-stopping logic in the markResult method, by
@@ -273,6 +299,268 @@ export class QueryInfo {
         this.lastWrite = void 0;
       }
     }
+  }
+
+  public markMutationResult<
+    TData,
+    TVariables extends OperationVariables,
+    TCache extends ApolloCache,
+  >(
+    mutation: {
+      mutationId: string;
+      result: FetchResult<TData>;
+      document: DocumentNode;
+      variables?: TVariables;
+      fetchPolicy?: MutationFetchPolicy;
+      errorPolicy: ErrorPolicy;
+      context?: DefaultContext;
+      updateQueries: UpdateQueries<TData>;
+      update?: MutationUpdaterFunction<TData, TVariables, TCache>;
+      awaitRefetchQueries?: boolean;
+      refetchQueries?: InternalRefetchQueriesInclude;
+      removeOptimistic?: string;
+      onQueryUpdated?: OnQueryUpdated<any>;
+      keepRootFields?: boolean;
+    },
+    cache = this.cache
+  ): Promise<FetchResult<TData>> {
+    let { result } = mutation;
+    const cacheWrites: Cache.WriteOptions[] = [];
+    const skipCache = mutation.fetchPolicy === "no-cache";
+
+    if (!skipCache && shouldWriteResult(result, mutation.errorPolicy)) {
+      if (!isExecutionPatchIncrementalResult(result)) {
+        cacheWrites.push({
+          result: result.data,
+          dataId: "ROOT_MUTATION",
+          query: mutation.document,
+          variables: mutation.variables,
+        });
+      }
+      if (
+        isExecutionPatchIncrementalResult(result) &&
+        isNonEmptyArray(result.incremental)
+      ) {
+        const diff = cache.diff<TData>({
+          id: "ROOT_MUTATION",
+          // The cache complains if passed a mutation where it expects a
+          // query, so we transform mutations and subscriptions to queries
+          // (only once, thanks to this.transformCache).
+          query: this.queryManager.getDocumentInfo(mutation.document).asQuery,
+          variables: mutation.variables,
+          optimistic: false,
+          returnPartialData: true,
+        });
+        let mergedData;
+        if (diff.result) {
+          mergedData = mergeIncrementalData(diff.result, result);
+        }
+        if (typeof mergedData !== "undefined") {
+          // cast the ExecutionPatchResult to FetchResult here since
+          // ExecutionPatchResult never has `data` when returned from the server
+          (result as FetchResult).data = mergedData;
+          cacheWrites.push({
+            result: mergedData,
+            dataId: "ROOT_MUTATION",
+            query: mutation.document,
+            variables: mutation.variables,
+          });
+        }
+      }
+
+      const { updateQueries } = mutation;
+      if (updateQueries) {
+        this.queryManager
+          .getObservableQueries("all")
+          .forEach((observableQuery) => {
+            const queryName = observableQuery && observableQuery.queryName;
+            if (
+              !queryName ||
+              !Object.hasOwnProperty.call(updateQueries, queryName)
+            ) {
+              return;
+            }
+            const updater = updateQueries[queryName];
+            const { query: document, variables } = observableQuery;
+
+            // Read the current query result from the store.
+            const { result: currentQueryResult, complete } =
+              observableQuery.getCacheDiff({ optimistic: false });
+
+            if (complete && currentQueryResult) {
+              // Run our reducer using the current query result and the mutation result.
+              const nextQueryResult = updater(currentQueryResult, {
+                mutationResult: result as FetchResult<Unmasked<TData>>,
+                queryName: (document && getOperationName(document)) || void 0,
+                queryVariables: variables!,
+              });
+
+              // Write the modified result back into the store if we got a new result.
+              if (nextQueryResult) {
+                cacheWrites.push({
+                  result: nextQueryResult,
+                  dataId: "ROOT_QUERY",
+                  query: document!,
+                  variables,
+                });
+              }
+            }
+          });
+      }
+    }
+
+    if (
+      cacheWrites.length > 0 ||
+      (mutation.refetchQueries || "").length > 0 ||
+      mutation.update ||
+      mutation.onQueryUpdated ||
+      mutation.removeOptimistic
+    ) {
+      const results: any[] = [];
+
+      this.queryManager
+        .refetchQueries({
+          updateCache: (cache) => {
+            if (!skipCache) {
+              cacheWrites.forEach((write) => cache.write(write));
+            }
+
+            // If the mutation has some writes associated with it then we need to
+            // apply those writes to the store by running this reducer again with
+            // a write action.
+            const { update } = mutation;
+            // Determine whether result is a SingleExecutionResult,
+            // or the final ExecutionPatchResult.
+            const isFinalResult =
+              !isExecutionPatchResult(result) ||
+              (isExecutionPatchIncrementalResult(result) && !result.hasNext);
+
+            if (update) {
+              if (!skipCache) {
+                // Re-read the ROOT_MUTATION data we just wrote into the cache
+                // (the first cache.write call in the cacheWrites.forEach loop
+                // above), so field read functions have a chance to run for
+                // fields within mutation result objects.
+                const diff = cache.diff<TData>({
+                  id: "ROOT_MUTATION",
+                  // The cache complains if passed a mutation where it expects a
+                  // query, so we transform mutations and subscriptions to queries
+                  // (only once, thanks to this.transformCache).
+                  query: this.queryManager.getDocumentInfo(mutation.document)
+                    .asQuery,
+                  variables: mutation.variables,
+                  optimistic: false,
+                  returnPartialData: true,
+                });
+
+                if (diff.complete) {
+                  result = { ...(result as FetchResult), data: diff.result };
+                  if ("incremental" in result) {
+                    delete result.incremental;
+                  }
+                  if ("hasNext" in result) {
+                    delete result.hasNext;
+                  }
+                }
+              }
+
+              // If we've received the whole response,
+              // either a SingleExecutionResult or the final ExecutionPatchResult,
+              // call the update function.
+              if (isFinalResult) {
+                update(
+                  cache as TCache,
+                  result as FetchResult<Unmasked<TData>>,
+                  {
+                    context: mutation.context,
+                    variables: mutation.variables,
+                  }
+                );
+              }
+            }
+
+            // TODO Do this with cache.evict({ id: 'ROOT_MUTATION' }) but make it
+            // shallow to allow rolling back optimistic evictions.
+            if (!skipCache && !mutation.keepRootFields && isFinalResult) {
+              cache.modify({
+                id: "ROOT_MUTATION",
+                fields(value, { fieldName, DELETE }) {
+                  return fieldName === "__typename" ? value : DELETE;
+                },
+              });
+            }
+          },
+
+          include: mutation.refetchQueries,
+
+          // Write the final mutation.result to the root layer of the cache.
+          optimistic: false,
+
+          // Remove the corresponding optimistic layer at the same time as we
+          // write the final non-optimistic result.
+          removeOptimistic: mutation.removeOptimistic,
+
+          // Let the caller of client.mutate optionally determine the refetching
+          // behavior for watched queries after the mutation.update function runs.
+          // If no onQueryUpdated function was provided for this mutation, pass
+          // null instead of undefined to disable the default refetching behavior.
+          onQueryUpdated: mutation.onQueryUpdated || null,
+        })
+        .forEach((result) => results.push(result));
+
+      if (mutation.awaitRefetchQueries || mutation.onQueryUpdated) {
+        // Returning a promise here makes the mutation await that promise, so we
+        // include results in that promise's work if awaitRefetchQueries or an
+        // onQueryUpdated function was specified.
+        return Promise.all(results).then(() => result);
+      }
+    }
+
+    return Promise.resolve(result);
+  }
+
+  public markMutationOptimistic<
+    TData,
+    TVariables extends OperationVariables,
+    TCache extends ApolloCache,
+  >(
+    optimisticResponse: any,
+    mutation: {
+      mutationId: string;
+      document: DocumentNode;
+      variables?: TVariables;
+      fetchPolicy?: MutationFetchPolicy;
+      errorPolicy: ErrorPolicy;
+      context?: DefaultContext;
+      updateQueries: UpdateQueries<TData>;
+      update?: MutationUpdaterFunction<TData, TVariables, TCache>;
+      keepRootFields?: boolean;
+    }
+  ) {
+    const data =
+      typeof optimisticResponse === "function" ?
+        optimisticResponse(mutation.variables, { IGNORE })
+      : optimisticResponse;
+
+    if (data === IGNORE) {
+      return false;
+    }
+
+    this.cache.recordOptimisticTransaction((cache) => {
+      try {
+        this.markMutationResult<TData, TVariables, TCache>(
+          {
+            ...mutation,
+            result: { data },
+          },
+          cache
+        );
+      } catch (error) {
+        invariant.error(error);
+      }
+    }, mutation.mutationId);
+
+    return true;
   }
 }
